fix(loongarch64): preserve full LSX task state

+++ crates/axcpu/src/loongarch64/context.rs
@@ -43,11 +43,11 @@ pub struct GeneralRegisters {
 }
 
 /// Floating-point registers of LoongArch64
-#[repr(C)]
+#[repr(C, align(16))]
 #[derive(Debug, Default, Clone, Copy)]
 pub struct FpuState {
-    /// Floating-point registers (f0-f31)
-    pub fp: [u64; 32],
+    /// Full LSX vector registers (vr0-vr31), including the aliased FP state.
+    pub fp: [u128; 32],
     /// Floating-point Condition Code register
     pub fcc: [u8; 8],
     /// Floating-point Control and Status register
